refactor: extrai funcoes de extracao de propriedade para um modulo proprio e refatora seus testes
