+++ src/features/common/shared/flags/modules/common/allowByUserId/index.ts
@@ -1,5 +1,6 @@
 import type { ReadonlyRequestCookies } from "next/dist/server/web/spec-extension/adapters/request-cookies";
-import { propertyFromCookie, propertyFromHeader } from "../logic";
+import { propertyFromHeader } from "../extractProperty/propertyFromHeader";
+import { propertyFromCookie } from "../extractProperty/propertyFromCookie";
 import type { UserId } from "./model";
 export { buildDecide } from "../logic";
